Student Record Management System Final Commit
It is a student record management system. It is developed in C++ In this You can create students and can keep the record of those students. The data is managed through file handling coded in C++. It is a very helpful system for the students who are learning C++ at the basic level and need to develop a system for their final evaluation. It covers all the concepts of language including OOP paradigm.